+++ index.html
@@ -8,6 +8,7 @@
 </head>
 <body>
   <canvas id="myCanvas"></canvas>
-  <script src="./js/Mizu.js"></script>
+  <script type="module" src="./js/Atom.js"></script>
+  <script type="module" src="./js/Mizu.js"></script>
 </body>
 </html>
